requirement.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
Django==5.2
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn
h11==0.14.0
h5py==3.13.0
haskellian==0.5.11
httpcore==1.0.8
httpx==0.28.1
idna==3.10
joblib==1.4.2
keras==3.9.2
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.3
namex==0.0.8
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
packaging==25.0
pandas==2.2.3
propcache==0.3.1
protobuf==5.29.4
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
pytz==2025.2
ramda==0.7.6
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==79.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
toolz==1.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
